restructured dataio and datasets to be part of the data package
