+++ docs/_templates/known_packages.rst
@@ -5,8 +5,8 @@
     For changes to take effect you might also have to run `make clean_all`
     afterwards.
 
-{% set known_packages=['glotaran.analysis', 'glotaran.dataio', 'glotaran.dataio.external_file_readers', 'glotaran.datasets', 'glotaran.model', 'glotaran.models', 'glotaran.models.damped_oscillation', 'glotaran.models.spectral_temporal', 'glotaran.parse', 'glotaran.plotting'] %}
-{% set child_modules=['glotaran.analysis', 'glotaran.analysis.dataset', 'glotaran.analysis.fitresult', 'glotaran.analysis.fitting', 'glotaran.analysis.grouping', 'glotaran.analysis.simulation', 'glotaran.analysis.variable_projection', 'glotaran.dataio', 'glotaran.dataio.chlorospec_format', 'glotaran.dataio.external_file_readers', 'glotaran.dataio.external_file_readers.sdt_reader', 'glotaran.dataio.file_readers', 'glotaran.dataio.legacy_readers', 'glotaran.dataio.mapper', 'glotaran.dataio.mlsd_file_format', 'glotaran.dataio.spectral_timetrace', 'glotaran.dataio.wavelength_time_explicit_file', 'glotaran.datasets', 'glotaran.datasets.dataset', 'glotaran.datasets.specialized_datasets', 'glotaran.datasets.spectral_temporal_dataset', 'glotaran.model', 'glotaran.model.base_model', 'glotaran.model.compartment_constraints', 'glotaran.model.dataset_descriptor', 'glotaran.model.model', 'glotaran.model.model_item', 'glotaran.model.model_item_validator', 'glotaran.model.parameter', 'glotaran.model.parameter_group', 'glotaran.models', 'glotaran.models.damped_oscillation', 'glotaran.models.damped_oscillation.doas_matrix', 'glotaran.models.damped_oscillation.doas_megacomplex', 'glotaran.models.damped_oscillation.doas_model', 'glotaran.models.damped_oscillation.doas_spectral_matrix', 'glotaran.models.damped_oscillation.oscillation', 'glotaran.models.spectral_temporal', 'glotaran.models.spectral_temporal.initial_concentration', 'glotaran.models.spectral_temporal.irf', 'glotaran.models.spectral_temporal.k_matrix', 'glotaran.models.spectral_temporal.kinetic_matrix', 'glotaran.models.spectral_temporal.kinetic_megacomplex', 'glotaran.models.spectral_temporal.kinetic_model', 'glotaran.models.spectral_temporal.kinetic_result', 'glotaran.models.spectral_temporal.spectral_matrix', 'glotaran.models.spectral_temporal.spectral_shape', 'glotaran.models.spectral_temporal.spectral_temporal_dataset_descriptor', 'glotaran.parse', 'glotaran.parse.parser', 'glotaran.parse.register', 'glotaran.plotting', 'glotaran.plotting.basic_plots', 'glotaran.plotting.glotaran_color_codes'] %}
+{% set known_packages=['glotaran.analysis', 'glotaran.data', 'glotaran.data.dataio', 'glotaran.data.datasets', 'glotaran.data.external_file_readers', 'glotaran.model', 'glotaran.models', 'glotaran.models.damped_oscillation', 'glotaran.models.spectral_temporal', 'glotaran.parse', 'glotaran.plotting'] %}
+{% set child_modules=['glotaran.analysis', 'glotaran.analysis.dataset', 'glotaran.analysis.fitresult', 'glotaran.analysis.fitting', 'glotaran.analysis.grouping', 'glotaran.analysis.simulation', 'glotaran.analysis.variable_projection', 'glotaran.data', 'glotaran.data.dataio', 'glotaran.data.dataio.chlorospec_format', 'glotaran.data.dataio.file_readers', 'glotaran.data.dataio.legacy_readers', 'glotaran.data.dataio.mapper', 'glotaran.data.dataio.mlsd_file_format', 'glotaran.data.dataio.spectral_timetrace', 'glotaran.data.dataio.wavelength_time_explicit_file', 'glotaran.data.datasets', 'glotaran.data.datasets.specialized_datasets', 'glotaran.data.datasets.spectral_temporal_dataset', 'glotaran.data.external_file_readers', 'glotaran.data.external_file_readers.sdt_reader', 'glotaran.model', 'glotaran.model.base_model', 'glotaran.model.compartment_constraints', 'glotaran.model.dataset', 'glotaran.model.dataset_descriptor', 'glotaran.model.model', 'glotaran.model.model_item', 'glotaran.model.model_item_validator', 'glotaran.model.parameter', 'glotaran.model.parameter_group', 'glotaran.models', 'glotaran.models.damped_oscillation', 'glotaran.models.damped_oscillation.doas_matrix', 'glotaran.models.damped_oscillation.doas_megacomplex', 'glotaran.models.damped_oscillation.doas_model', 'glotaran.models.damped_oscillation.doas_spectral_matrix', 'glotaran.models.damped_oscillation.oscillation', 'glotaran.models.spectral_temporal', 'glotaran.models.spectral_temporal.initial_concentration', 'glotaran.models.spectral_temporal.irf', 'glotaran.models.spectral_temporal.k_matrix', 'glotaran.models.spectral_temporal.kinetic_matrix', 'glotaran.models.spectral_temporal.kinetic_megacomplex', 'glotaran.models.spectral_temporal.kinetic_model', 'glotaran.models.spectral_temporal.kinetic_result', 'glotaran.models.spectral_temporal.spectral_matrix', 'glotaran.models.spectral_temporal.spectral_shape', 'glotaran.models.spectral_temporal.spectral_temporal_dataset_descriptor', 'glotaran.parse', 'glotaran.parse.parser', 'glotaran.parse.register', 'glotaran.plotting', 'glotaran.plotting.basic_plots', 'glotaran.plotting.glotaran_color_codes'] %}
 {% block module %}
 
 {% endblock %}
