+++ glotaran/data/dataio/chlorospec_format.py
@@ -1,4 +1,4 @@
-from glotaran.datasets.spectral_temporal_dataset import SpectralTemporalDataset
+from glotaran.data.datasets.spectral_temporal_dataset import SpectralTemporalDataset
 from natsort import natsorted
 from glob import glob
 import os
