+++ glotaran/data/dataio/file_readers.py
@@ -6,11 +6,11 @@ import warnings
 
 import numpy as np
 import pandas as pd
-from .external_file_readers.sdt_reader import SdtFile
+from glotaran.data.external_file_readers.sdt_reader import SdtFile
 from .mapper import get_pixel_map
-from ..datasets.spectral_temporal_dataset import SpectralTemporalDataset
-from ..datasets.specialized_datasets import FLIMDataset
-from ..datasets.dataset import DimensionalityError
+from glotaran.data.datasets.spectral_temporal_dataset import SpectralTemporalDataset
+from glotaran.data.datasets.specialized_datasets import FLIMDataset
+from glotaran.model.dataset import DimensionalityError
 
 
 def dataframe_to_SpectralTemporalDataset(input_dataframe: pd.DataFrame,
