+++ glotaran/data/dataio/test/data_files/FLIM.sdt
