+++ glotaran/data/dataio/test/data_files/FLIM_legacy_sum_map.ssv
