+++ glotaran/data/dataio/test/data_files/FLIM_legacy_trace.ssv
