+++ glotaran/data/dataio/test/test_legacy_readers.py
